requirements.txt
django
django-ckeditor
django-mathfilters
Pillow
bcrypt
django[argon2]
#scrypt
djangorestframework
pdfkit
mlxtend
pandas
numpy
openpyxl
humanize
feedparser
matplotlib
seaborn